requirements.txt
attrs==18.2.0
Automat==0.7.0
certifi>=2018.8.24
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
PyHamcrest==1.9.0
requests>=2.19.1
requests-threads>=0.1.1
six==1.11.0
Twisted>=18.7.0
urllib3==1.23
zope.interface>=4.5.0
privex-loghelper
privex-helpers>=1.1.0